1) Corrected the parameter type for setTemp method of main class 2)Implemented unit test cases for validate method of WeatherResponseResource  class. 3)Corrected the validate method of WeatherResponseResource class after checking unit tests are being fialed.

+++ src/main/java/com/example/Weather/SpringWeather/external/Main.java
@@ -1,6 +1,5 @@
 package com.example.Weather.SpringWeather.external;
 
-import org.springframework.stereotype.Component;
 
 public class Main {
 	
@@ -27,7 +26,7 @@ public class Main {
 		return temp;
 	}
 
-	public void setTemp(int temp) {
+	public void setTemp(double temp) {
 		this.temp = temp;
 	}
